Add init. Update d3-line-graph.
init is a starter template for meteor projects. I'm not sure it can be used since each Meteor project requires a unique ID.

d3-line-graph has been updated with a revised README and the ability to display data from the database as a table. The d3 chart isn't yet working.